acceptance: inline a smarter postgres-test image
Ditch cockroachdb/postgres-test for a new image, cockroachdb/acceptance.
This new image has a few perks:

   * It avoids dependencies on external package managers.

   * It's smarter about ordering commands in the Dockerfile to minimize cache
     invalidation when iterating on the Dockerfile itself.

   * It lives in the cockroachdb repository (pkg/acceptance/testdata,
     specifically) instead of its own repository, which means we no
     longer need to duplicate dependency information in two places.
     This also generally decreases friction.

     The other user of the old image is examples-orms, which is slated
     to be moved into cockroachdb/cockroach eventually. Even if that
     doesn't happen, we're not making the separation any worse.

Fixes #19841.
