+++ pkg/acceptance/node_test.go
@@ -38,6 +38,6 @@ cd /testdata/node
 
 export SHOULD_FAIL=%v
 # Get access to globally installed node modules.
-export NODE_PATH=$NODE_PATH:/usr/lib/node_modules/
-mocha .
+export NODE_PATH=$NODE_PATH:/usr/lib/node
+/usr/lib/node/.bin/mocha .
 `
