+++ pkg/acceptance/testdata/csharp/.gitignore
@@ -1,0 +1,3 @@
+bin
+obj
+node.pfx
